Preping for sockets
Refactored how authentication works so socket requests can be authenticated. Now instead of using middleware any function
that needs to authenticate will get the token and pass it into the auth util function which will return a user if the token is valid.
I also added the token to cookies so it would be easy to grab and send with socket data. Some initial socket set up was also done but not much.