Set up sockets
Now using sockets to create/delete pins and create comments. This was done so that as users interact with the map other users clients will be updated.
Currently the front end emits an event with the affected pin, the back end processes the pin and then sends the new pin to every client.
I now need to set up the font end to update state with the new pins.